changes for the pdf report new updated new one

+++ controllers/ReportController.js
@@ -1063,29 +1063,35 @@ exports.createCaseReport = [
 
 						}
 
-						fs.unlink(process.cwd()+"/uploads/delete_created_files/case_report_"+caseId+".pdf", (err) => {
-							if (err) {
-							  // Handle the error
-							} else {
-							  // File successfully deleted
-							}
-						});
+						if(fs.exists(process.cwd()+"/uploads/delete_created_files/case_report_"+caseId+".pdf")) {
+							fs.unlink(process.cwd()+"/uploads/delete_created_files/case_report_"+caseId+".pdf", (err) => {
+								if (err) {
+								console.log("error in dleeteing file case report "+err);
+								} else {
+									console.log("success in dleeteing file case report ");
+								}
+							});
+						}
 
-						fs.unlink(process.cwd()+"/uploads/delete_created_files/ecg_report_" + caseId + ".pdf", (err) => {
-							if (err) {
-							  // Handle the error
-							} else {
-							  // File successfully deleted
-							}
-						});
+						if(fs.exists(process.cwd()+"/uploads/delete_created_files/ecg_report_" + caseId + ".pdf")) {
+							fs.unlink(process.cwd()+"/uploads/delete_created_files/ecg_report_" + caseId + ".pdf", (err) => {
+								if (err) {
+									console.log("error in dleeteing file egc report "+err);
+								} else {
+									console.log("success in deleting file edg report ");
+								}
+							});
+						}
 
-						fs.unlink(process.cwd()+"/uploads/delete_created_files/"+"sh_file_" + caseId + ".sh", (err) => {
-							if (err) {
-							  // Handle the error
-							} else {
-							  // File successfully deleted
-							}
-						});
+						if(fs.exists(process.cwd()+"/uploads/delete_created_files/"+"sh_file_" + caseId + ".sh")) {
+							fs.unlink(process.cwd()+"/uploads/delete_created_files/"+"sh_file_" + caseId + ".sh", (err) => {
+								if (err) {
+									console.log("error in dleeteing file sh report "+err);
+								} else {
+									console.log("success in deleting file sh report ");
+								}
+							});
+						}
